Organize evals into lib/ and suites/, with one run command
Separates the harness from the evals it runs. lib/ holds the provider,
the fixture builder and the runner; suites/ holds one directory per eval,
each self-contained with its config, prompt, cases and fixture. Adding an
eval is now copying a directory and editing four files — nothing in lib/
or package.json changes, since both the runner and validate glob
suites/*/promptfooconfig.yaml.

Keeps a single `eval` command that behaves the way jest and playwright do:
no argument runs every suite, a leading path runs one, and flags pass
through either way.

That needs lib/run.sh rather than an npm script. npm appends arguments to
the script string instead of passing them as "$@", so a loop ending in
`done` is a syntax error the moment anything is appended, and a script
ending in `--config` breaks when nothing is. Invoking `sh -c … --` gives
the arguments back as positional parameters, and the branch on whether the
first one looks like a path is what restores the familiar behaviour.

Passing several configs to promptfoo is not an alternative: it merges them
into a cross-product rather than running them in sequence. Two configs
with one prompt and one test each produce four cases, with each suite's
assertions applied to the other's output. Both the runner and validate
therefore loop.

fixture_module now resolves against the declaring config's basePath rather
than the harness directory, so a suite refers to its fixture by bare
filename and stays copy-pasteable.

Also adopts promptfoo's own config-schema comment for editor autocomplete.
