+++ package.json
@@ -185,8 +185,6 @@
 		"test:unit:update": "npm run --workspace @wordpress/unit-tests test:unit -- --updateSnapshot",
 		"test:unit:watch": "npm run --workspace @wordpress/unit-tests test:unit:watch --",
 		"test:agent-evals": "npm --prefix test/ai-development/evals run eval --",
-		"test:agent-evals:claude": "npm --prefix test/ai-development/evals run eval:claude",
-		"test:agent-evals:codex": "npm --prefix test/ai-development/evals run eval:codex",
 		"wp-env": "wp-env",
 		"wp-env-test": "wp-env --config .wp-env.test.json"
 	},
